+++ CHANGELOG.md
@@ -6,9 +6,14 @@
 
 ### Fixed
 - Fixed a bug that blocked arbitrum mainnet wallets from being created
+### Added
 - Add `deployToken` method to `WalletAddress` and `Wallet` to deploy an ERC20, updated `SmartContract` class to support deployment and fetching contract details
 - Add SOL asset support
+- Fix a bug where large numbers were being returned in scientific notation
 
+### Breaking
+- `Asset#toAtomicAmount` now returns a BigInt instead of a Decimal
+  
 ## [0.5.0] - 2024-09-11
 
 - Add Arbitrum-Mainnet support for Native transfers.